requirements.txt
# web/api
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
httpx>=0.27.0
python-dotenv>=1.0.1

# PDF & OCR (PyMuPDF 제거)
pypdf==5.1.0
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0

# NLP 기초
regex==2024.9.11
nltk==3.9.1

# LLM (Gemini)
google-generativeai>=0.7.0

# DB
SQLAlchemy>=2.0.32
psycopg[binary]>=3.2.1

# 벡터 검색
scikit-learn==1.5.2    # (옵션) cosine 거리 계산/파이프라인
